feat(server): detach process

+++ BeDbg/Models/DebuggingProcess.cs
@@ -4,5 +4,5 @@ public class DebuggingProcess
 {
 	public DateTime AttachTime { get; set; }
 	public int Id { get; set; }
-	public IntPtr Handle { get; set; }
+	public long Handle { get; set; }
 }
